+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the Create Flutter Widgets and Classes extension will be 
 ## [1.1.0]
 
 - Added the code action `Wrap with Observer`.
+- Added the code action `Wrap with Stack`.
+- Added the code action `Wrap with Expanded`.
+- Added the code action `Wrap with Positioned`.
+- Added the code action `Wrap with Form`.
+- Added the code action `Wrap with SafeArea`.
 
 ## [1.0.1]